Use of UTF8 tokenizer and various changes
Use of a tokenizer.  Generator have been adapted to work with tuples of strings instead of pure strings.
Use of "lexicon" instead of "dictionary" to denote weighted name lists in the source.
Use of "\x02" and "\x03" as start and end of word delimiter.
Without using the "--number" flag explicitly in command line, the GUI doesn't auto-generate n words
Proper `unset()` method for class `Selector.Arguments`.